LUT-29875 : check task access rights feature

+++ src/java/fr/paris/lutece/plugins/taskstack/service/TaskService.java
@@ -33,6 +33,15 @@
  */
 package fr.paris.lutece.plugins.taskstack.service;
 
+import java.sql.Timestamp;
+import java.util.ArrayList;
+import java.util.Date;
+import java.util.HashMap;
+import java.util.List;
+import java.util.Map;
+import java.util.UUID;
+import java.util.stream.Collectors;
+
 import fr.paris.lutece.plugins.taskstack.business.task.Task;
 import fr.paris.lutece.plugins.taskstack.business.task.TaskChange;
 import fr.paris.lutece.plugins.taskstack.business.task.TaskChangeHome;
@@ -51,55 +60,67 @@ import fr.paris.lutece.portal.service.spring.SpringContextService;
 import fr.paris.lutece.portal.service.util.AppPropertiesService;
 import fr.paris.lutece.util.sql.TransactionManager;
 
-import java.sql.Timestamp;
-import java.util.ArrayList;
-import java.util.Date;
-import java.util.HashMap;
-import java.util.List;
-import java.util.Map;
-import java.util.Objects;
-import java.util.UUID;
-import java.util.stream.Collectors;
-
 public class TaskService
 {
-    private static final int PROPERTY_MAX_NB_TASK_RETURNED = AppPropertiesService.getPropertyInt("taskstack.search.maxNbTaskReturned", 0);
-
+    private static final int PROPERTY_MAX_NB_TASK_RETURNED = AppPropertiesService.getPropertyInt( "taskstack.search.maxNbTaskReturned", 0 );
 
     private static TaskService _instance;
+    private static Map<String, ITaskManagement> taskManagementBeans;
 
+    /**
+     * get instance of taskService
+     * 
+     * @return the instance
+     */
     public static TaskService instance( )
     {
         if ( _instance == null )
         {
             _instance = new TaskService( );
+
+            Map<String, ITaskManagement> listOfBeans = SpringContextService.getContext( ).getBeansOfType( ITaskManagement.class );
+            taskManagementBeans = new HashMap<>( );
+            listOfBeans.forEach( ( key, taskManager ) -> {
+                taskManagementBeans.put( taskManager.getTaskType( ), taskManager );
+            } );
         }
         return _instance;
     }
 
+    /**
+     * Private Constructor
+     */
     private TaskService( )
     {
     }
 
+    /**
+     * Create task
+     * 
+     * @param taskDto
+     * @param author
+     * @param clientCode
+     * @return the TaskDto of the created task
+     * @throws TaskStackException
+     */
     public TaskDto createTask( final TaskDto taskDto, final RequestAuthor author, final String clientCode ) throws TaskStackException
     {
-        final Map<String, ITaskManagement> taskManagementBeans = SpringContextService.getContext( ).getBeansOfType( ITaskManagement.class );
-        final ITaskManagement taskManagement = taskManagementBeans.values( ).stream( ).filter( t -> Objects.equals( t.getTaskType( ), taskDto.getTaskType( ) ) )
-                .findFirst( ).orElse( null );
+        final ITaskManagement taskManager = taskManagementBeans.get( taskDto.getTaskType( ) );
+
         TransactionManager.beginTransaction( null );
         try
         {
             boolean validationError = false;
-            Map<String,String> errMap = new HashMap<>( );
+            Map<String, String> errMap = new HashMap<>( );
             taskDto.setTaskCode( UUID.randomUUID( ).toString( ) );
             taskDto.setLastUpdateClientCode( clientCode );
             taskDto.setTaskStatus( TaskStatusType.TODO );
 
-            if ( taskManagement != null )
+            if ( taskManager != null )
             {
                 try
                 {
-                    taskManagement.doBefore( taskDto );
+                    taskManager.doBefore( taskDto );
                 }
                 catch( final TaskValidationException e )
                 {
@@ -124,13 +145,13 @@ public class TaskService
 
             TaskChangeHome.create( taskChange );
 
-            if ( taskManagement != null )
+            if ( taskManager != null )
             {
-                taskManagement.doAfter( taskDto );
+                taskManager.doAfter( taskDto );
             }
 
             TransactionManager.commitTransaction( null );
-            
+
             return taskDto;
         }
         catch( final Exception e )
@@ -140,6 +161,15 @@ public class TaskService
         }
     }
 
+    /**
+     * Update status
+     * 
+     * @param strTaskCode
+     * @param newStatus
+     * @param author
+     * @param clientCode
+     * @throws TaskStackException
+     */
     public void updateTaskStatus( final String strTaskCode, final TaskStatusType newStatus, final RequestAuthor author, final String clientCode )
             throws TaskStackException
     {
@@ -149,16 +179,19 @@ public class TaskService
             throw new TaskNotFoundException( "Could not find task with code " + strTaskCode );
         }
 
-        final ITaskManagement taskManagement = SpringContextService.getBeansOfType( ITaskManagement.class ).stream( )
-                .filter( t -> Objects.equals( t.getTaskType( ), existingTask.getTaskType( ) ) ).findFirst( ).orElse( null );
+        final ITaskManagement taskManager = taskManagementBeans.get( existingTask.getTaskType( ) );
+
         final TaskDto existingTaskDto = DtoMapper.toTaskDto( existingTask );
+
+        checkAccess( existingTaskDto, TaskChangeType.UPDATED );
+
         existingTaskDto.setTaskStatus( newStatus );
         TransactionManager.beginTransaction( null );
         try
         {
-            if ( taskManagement != null )
+            if ( taskManager != null )
             {
-                taskManagement.doBefore( existingTaskDto );
+        	taskManager.doBefore( existingTaskDto );
             }
             final Timestamp updateStatusTime = TaskHome.updateStatus( strTaskCode, newStatus, clientCode );
 
@@ -171,9 +204,9 @@ public class TaskService
             taskChange.setAuthorName( author.getName( ) );
             taskChange.setTaskChangeDate( updateStatusTime );
             TaskChangeHome.create( taskChange );
-            if ( taskManagement != null )
+            if ( taskManager != null )
             {
-                taskManagement.doAfter( existingTaskDto );
+        	taskManager.doAfter( existingTaskDto );
             }
 
             TransactionManager.commitTransaction( null );
@@ -185,6 +218,13 @@ public class TaskService
         }
     }
 
+    /**
+     * Get a task
+     * 
+     * @param strTaskCode
+     * @return the taskDto
+     * @throws TaskStackException
+     */
     public TaskDto getTask( final String strTaskCode ) throws TaskStackException
     {
         final Task existingTask = TaskHome.get( strTaskCode );
@@ -195,10 +235,43 @@ public class TaskService
         }
 
         final TaskDto taskDto = DtoMapper.toTaskDto( existingTask );
+
+        checkAccess( taskDto, TaskChangeType.READ );
+
         taskDto.getTaskChanges( ).addAll( this.getTaskHistory( taskDto.getTaskCode( ) ) );
         return taskDto;
     }
 
+    /**
+     * check access
+     * 
+     * @param taskDto
+     * @throws TaskNotFoundException
+     */
+    private void checkAccess( TaskDto taskDto, TaskChangeType type ) throws TaskNotFoundException
+    {
+        final ITaskManagement taskManager = taskManagementBeans.get( taskDto.getTaskType( ) );
+        if ( taskManager != null )
+        {
+            try
+            {
+                taskManager.checkAccess( taskDto, type );
+            }
+            catch( TaskValidationException e )
+            {
+                throw new TaskNotFoundException( "Access Denied" );
+            }
+        }
+    }
+
+    /**
+     * get the list of tasks for a given resource
+     * 
+     * @param strResourceId
+     * @param strResourceType
+     * @return the task list
+     * @throws TaskStackException
+     */
     public List<TaskDto> getTasks( final String strResourceId, final String strResourceType ) throws TaskStackException
     {
         final List<Task> tasks = TaskHome.get( strResourceId, strResourceType );
@@ -209,23 +282,107 @@ public class TaskService
             for ( final Task task : tasks )
             {
                 final TaskDto taskDto = DtoMapper.toTaskDto( task );
-                taskDto.getTaskChanges( ).addAll( this.getTaskHistory( task.getTaskCode( ) ) );
-                taskDtos.add( taskDto );
+                try
+                {
+                    checkAccess( taskDto, TaskChangeType.READ );
+                    taskDto.getTaskChanges( ).addAll( this.getTaskHistory( task.getTaskCode( ) ) );
+                    taskDtos.add( taskDto );
+                }
+                catch( TaskNotFoundException e )
+                {
+                    // do not add to list
+                }
+
             }
         }
 
         return taskDtos;
     }
 
-    public List<TaskDto> search(final String _strTaskCode, final String _strResourceId, final String _strResourceType, final String _strTaskType, final Date creationDate, final Date lastUpdatedate, final String strLastUpdateClientCode, final List<TaskStatusType> _enumTaskStatus, final Integer _nNbDaysSinceCreated,
-                                final CreationDateOrdering creationDateOrdering, final Map<String, String> metadata, final int max ) throws TaskStackException
+    /**
+     * Search tasks with criterias
+     * 
+     * @param _strTaskCode
+     * @param _strResourceId
+     * @param _strResourceType
+     * @param _strTaskType
+     * @param creationDate
+     * @param lastUpdatedate
+     * @param strLastUpdateClientCode
+     * @param _enumTaskStatus
+     * @param _nNbDaysSinceCreated
+     * @param creationDateOrdering
+     * @param metadata
+     * @param max
+     * @return the task list
+     * @throws TaskStackException
+     */
+    public List<TaskDto> search( final String _strTaskCode, final String _strResourceId, final String _strResourceType, final String _strTaskType,
+            final Date creationDate, final Date lastUpdatedate, final String strLastUpdateClientCode, final List<TaskStatusType> _enumTaskStatus,
+            final Integer _nNbDaysSinceCreated, final CreationDateOrdering creationDateOrdering, final Map<String, String> metadata, final int max )
+            throws TaskStackException
+    {
+        int nMaxNbIdentityReturned = ( max > 0 ) ? max : PROPERTY_MAX_NB_TASK_RETURNED;
+        List<TaskDto> taskList = new ArrayList<>( );
+
+        try
+        {
+            final List<Task> result = TaskHome.search( _strTaskCode, _strResourceId, _strResourceType, _strTaskType, creationDate, lastUpdatedate,
+                    strLastUpdateClientCode, _enumTaskStatus, _nNbDaysSinceCreated, creationDateOrdering, nMaxNbIdentityReturned, metadata );
+
+            for ( Task t : result )
+            {
+                try
+                {
+                    final TaskDto taskDto = DtoMapper.toTaskDto( t );
+
+                    checkAccess( taskDto, TaskChangeType.READ );
+                    taskDto.getTaskChanges( ).addAll( this.getTaskHistory( taskDto.getTaskCode( ) ) );
+                    taskList.add( taskDto );
+                }
+                catch( TaskNotFoundException e )
+                {
+                    // do not add to list
+                }
+            }
+        }
+        catch( final Exception e )
+        {
+            throw new TaskStackException( "An error occurred during task search", e );
+        }
+
+        return taskList;
+
+    }
+
+    /**
+     * Search a list of Task Ids
+     * 
+     * @param _strTaskCode
+     * @param _strResourceId
+     * @param _strResourceType
+     * @param _strTaskType
+     * @param creationDate
+     * @param lastUpdatedate
+     * @param strLastUpdateClientCode
+     * @param _enumTaskStatus
+     * @param _nNbDaysSinceCreated
+     * @param creationDateOrdering
+     * @param metadata
+     * @param max
+     * @return the list of Ids of the tasks
+     * @throws TaskStackException
+     */
+    public List<Integer> searchId( final String _strTaskCode, final String _strResourceId, final String _strResourceType, final String _strTaskType,
+            final Date creationDate, final Date lastUpdatedate, final String strLastUpdateClientCode, final List<TaskStatusType> _enumTaskStatus,
+            final Integer _nNbDaysSinceCreated, final CreationDateOrdering creationDateOrdering, final Map<String, String> metadata, final int max )
+            throws TaskStackException
     {
         int nMaxNbIdentityReturned = ( max > 0 ) ? max : PROPERTY_MAX_NB_TASK_RETURNED;
         try
         {
-            final List<Task> search = TaskHome.search( _strTaskCode, _strResourceId, _strResourceType, _strTaskType, creationDate, lastUpdatedate, strLastUpdateClientCode, _enumTaskStatus, _nNbDaysSinceCreated, creationDateOrdering, nMaxNbIdentityReturned, metadata );
-            return search.stream( ).map( DtoMapper::toTaskDto )
-                    .peek( taskDto -> taskDto.getTaskChanges( ).addAll( this.getTaskHistory( taskDto.getTaskCode( ) ) ) ).collect( Collectors.toList( ) );
+            return TaskHome.searchId( _strTaskCode, _strResourceId, _strResourceType, _strTaskType, creationDate, lastUpdatedate, strLastUpdateClientCode,
+                    _enumTaskStatus, _nNbDaysSinceCreated, creationDateOrdering, nMaxNbIdentityReturned, metadata );
         }
         catch( final Exception e )
         {
@@ -233,37 +390,53 @@ public class TaskService
         }
     }
 
-    public List<Integer> searchId(final String _strTaskCode, final String _strResourceId, final String _strResourceType, final String _strTaskType, final Date creationDate, final Date lastUpdatedate, final String strLastUpdateClientCode, final List<TaskStatusType> _enumTaskStatus, final Integer _nNbDaysSinceCreated,
-                                final CreationDateOrdering creationDateOrdering, final Map<String, String> metadata, final int max ) throws TaskStackException
+    /**
+     * get a list of tasks from a list of Ids
+     * 
+     * @param listId
+     * @return the task list
+     * @throws TaskStackException
+     */
+    public List<TaskDto> getTasksListByIds( List<Integer> listId ) throws TaskStackException
     {
-        int nMaxNbIdentityReturned = ( max > 0 ) ? max : PROPERTY_MAX_NB_TASK_RETURNED;
-        try
-        {
-            return TaskHome.searchId( _strTaskCode, _strResourceId, _strResourceType, _strTaskType, creationDate, lastUpdatedate, strLastUpdateClientCode, _enumTaskStatus, _nNbDaysSinceCreated, creationDateOrdering, nMaxNbIdentityReturned, metadata );
-        }
-        catch( final Exception e )
-        {
-            throw new TaskStackException( "An error occurred during task search", e );
-        }
+        final List<Task> search = TaskHome.getTasksListByIds( listId );
+        return search.stream( ).map( DtoMapper::toTaskDto ).peek( taskDto -> taskDto.getTaskChanges( ).addAll( this.getTaskHistory( taskDto.getTaskCode( ) ) ) )
+                .collect( Collectors.toList( ) );
     }
 
-    public List<TaskDto> getTasksListByIds(List<Integer> listId) throws TaskStackException
-    {
-        final List<Task> search = TaskHome.getTasksListByIds(listId);
-        return search.stream( ).map( DtoMapper::toTaskDto )
-                .peek( taskDto -> taskDto.getTaskChanges( ).addAll( this.getTaskHistory( taskDto.getTaskCode( ) ) ) ).collect( Collectors.toList( ) );
-    }
-
+    /**
+     * Delete task
+     * 
+     * @param task
+     * @throws TaskStackException
+     */
     public void deleteTask( final Task task ) throws TaskStackException
     {
         if ( task == null || task.getId( ) == null )
         {
             throw new TaskStackException( "Could not delete null task " );
         }
-        TaskChangeHome.deleteAllByTaskId( task.getId( ) );
-        TaskHome.delete( task.getId( ) );
+
+        try
+        {
+            checkAccess( DtoMapper.toTaskDto( task ), TaskChangeType.DELETED );
+
+            TaskChangeHome.deleteAllByTaskId( task.getId( ) );
+            TaskHome.delete( task.getId( ) );
+        }
+        catch( TaskNotFoundException e )
+        {
+            // do not delete
+        }
+
     }
 
+    /**
+     * get task history
+     * 
+     * @param strTaskCode
+     * @return the history items list
+     */
     private List<TaskChangeDto> getTaskHistory( final String strTaskCode )
     {
         final List<TaskChange> history = TaskChangeHome.getHistory( strTaskCode );
